Moved poller test to own file.
